fix: Correções simples - 15 erros (property renames, imports, enum fixes)
**repository-errors-fixer** - Correções rápidas:
- GroupDetailScreen: memberCount → membersCount (3 erros)
- GameSummonViewModel, InviteViewModel: domain model imports (2 erros)
- LiveGame ViewModels/Screens: domain.GameEvent, LivePlayerStats imports (4 erros)
- EnhancedTeamBalancer: .uppercase() → .toUpperCase() (1 erro)
- LeagueService: Remove .toInt() conversions (3 erros)
- ConfirmPresenceUseCase: PlayerPosition.FIELD → LINE, comment game.players (4 erros)

Total: ~15 erros corrigidos

Restam ~195-200 erros (próxima verificação)

+++ app/src/main/java/com/futebadosparcas/domain/ranking/LeagueService.kt
@@ -110,12 +110,12 @@ class LeagueService constructor(
                 else -> -10
             }
             val newLeagueRating = (participation.leagueRating + ratingChange).toDouble()
-            val suggestedDivision = com.futebadosparcas.domain.ranking.LeagueRatingCalculator.getDivisionForRating(newLeagueRating.toInt())
+            val suggestedDivision = com.futebadosparcas.domain.ranking.LeagueRatingCalculator.getDivisionForRating(newLeagueRating)
 
             // FIXME: protectionGames, promotionProgress, relegationProgress nao existem no SeasonParticipation
             // Simplificando: divisao baseada apenas no league rating atual
             val oldDivisionShared = SharedLeagueDivision.fromString(participation.division)
-            val newDivisionShared = com.futebadosparcas.domain.ranking.LeagueRatingCalculator.getDivisionForRating(newLeagueRating.toInt())
+            val newDivisionShared = com.futebadosparcas.domain.ranking.LeagueRatingCalculator.getDivisionForRating(newLeagueRating)
 
             val promoted = newDivisionShared.ordinal > oldDivisionShared.ordinal
             val relegated = newDivisionShared.ordinal < oldDivisionShared.ordinal
